feat: allow to run loki with existing headless chrome docker container
This adds a new target for running storybook in chrome on an existing
docker container. The target isn't exported directly, but as a variant
of the chromeDockerTarget to make it possible to let Loki run docker
on a local machine while enabling the use of an existing container with
a flag on the CI. This way the same target config can be used on local
and on the CI.
The biggest challenge with this change was to provide access to local
files as we cannot assume that the external headless chrome docker
container has access to those. Instead Loki will run a small static
server to serve the file and direct the headless chrome towards this
page.
Two new packages are added: bluebird for better promise handling,
especially timeouts, and serve-handler to serve the static directory.
All changes are behind the chromeDockerUseExisting flag which is false
by default, so there should be no breaking changes.