Rename `chromeTolerance` to `tolerance` and pass to all targets
`tolerance` is respected by `gm` and `looks-same` yet we weren't passing the option for the iOS simulator and Android emulator. This patch rectifies that.

I've removed the `tolerance / 100` in `diffing/gm` for consistency between the two engines so that the may user decide.

Lastly, I've added a section on tolerance to clear up any ambiguity which should resolve https://github.com/oblador/loki/issues/113