add `chrome.autoExpandHeight` option
which when enabled will extend the viewport height if necessary so that the area that is being screenshotted does not go below the fold.

With chrome 76 (and probably earlier) any area below the fold shows up as white.

It looks like a while back the screenshot would still work even when the area went below the fold, because with the default 'yukinying/chrome-headless' image (that hasn't been updated in a while) anything below the fold is still captured.